add the src folder
This commit includes:
- index.js: [main file] extends the chai package, adding the
.matchSnapshot and .matchSpecificSnapshot to its expect matchers.
- matchSnapshot.js: module with the matchSnapshot function, usded for
both .matchSnapshot and .matchSpecificSnapshot matchers.
- checkCI.js: checks if the tests are running in a CI system. Note that,
the snapshots must already be included in your project if you want to
run the tests in a CI system.
- getExistingSnaps.js: import the already existing snapshots.
- snapshotDirConstructor.js: responsible for creating the path of the
directory where the snapshot file (.snap) will be. By default, the .snap
file is saved within the __snapshots__ folder (located at the same level
of the test file). The matchSpecificSnapshot includes the possibility
for the user to create subfolders inside the __snapshots__ folder, or
even choose a totally different folder location by passing the desired
absolute path.
- testNameConstructor.js: responsible for creating each snapshot name.
By default, each snapshot has the name of the test where the matcher is
used (in this case, the snapshot name is the concatenation of the 'it'
title with the titles of the "describe" parents). An optional hint can
be badded at the end of each snapshot name. The matchSpecificSnapshot
includes the possibility for the user to provide an entire new snap
name.
- writeSnapshots.js: responsible to write the new snapshots in the .snap
file.